LPS-71280 Moves ExportImportConfigurationTestUtil
