+++ modules/apps/web-experience/export-import/export-import-test/src/testIntegration/java/com/liferay/exportimport/trash/test/util/ExportImportConfigurationTestUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.layoutsadmin.util.test;
+package com.liferay.exportimport.trash.test.util;
 
 import com.liferay.exportimport.kernel.configuration.ExportImportConfigurationSettingsMapFactory;
 import com.liferay.exportimport.kernel.model.ExportImportConfiguration;
